CLN: Improve QueryBuilder and SelectBuilder API
